Convert next & previous component to gem-based
- Change the class from `pub-c` to `gem-c`
- Fix CSS includes
- Rename files
- Make tests work
- Update CHANGELOG
