+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   tags, OpenGraph tags and Twitter cards (#335).
 * Move the Title component from static (PR #324)
 * Move the Lead paragraph component from static (PR #325)
+* Move the Next and previous component from static (PR #329)
 * Add a Phase banner component to replace the Alpha/Beta banners in Static (PR #333)
 
 ### Upgrade instructions
@@ -21,6 +22,7 @@
 Replace the component by running the following:
 
 ```
+find app/views -type f -print0 | xargs -0 sed -i '' 's/govuk_component\/previous_and_next_navigation/govuk_publishing_components\/components\/previous_and_next_navigation/g'
 find app/views -type f -print0 | xargs -0 sed -i '' 's/govuk_component\/title/govuk_publishing_components\/components\/title/g'
 find app/views -type f -print0 | xargs -0 sed -i '' 's/govuk_component\/lead_paragraph/govuk_publishing_components\/components\/lead_paragraph/g'
 ```
