+++ app/assets/stylesheets/govuk_publishing_components/_all_components.scss
@@ -22,6 +22,7 @@
 @import "components/label";
 @import "components/lead-paragraph";
 @import "components/phase-banner";
+@import "components/previous-and-next-navigation";
 @import "components/radio";
 @import "components/related-navigation";
 @import "components/search";
