+++ spec/components/previous_and_next_navigation_spec.rb
@@ -1,0 +1,74 @@
+require 'rails_helper'
+
+describe "Previous and next navigation", type: :view do
+  def component_name
+    "previous_and_next_navigation"
+  end
+
+  it "renders if no parameters" do
+    assert_empty render_component({})
+  end
+
+  it "renders previous pagination" do
+    render_component(previous_page: {
+      url: "previous-page",
+      title: "Previous page",
+      label: "1 of 3"
+    })
+
+    assert_select ".gem-c-pagination[role='navigation']"
+    assert_select ".gem-c-pagination__link-title", text: "Previous page"
+    assert_select ".gem-c-pagination__link-label", text: "1 of 3"
+    assert_link("previous-page")
+  end
+
+  it "render pagination" do
+    render_component(next_page: {
+      url: "next-page",
+      title: "Next page",
+      label: "2 of 3"
+    })
+
+    assert_select ".gem-c-pagination__link-title", text: "Next page"
+    assert_select ".gem-c-pagination__link-label", text: "2 of 3"
+    assert_link("next-page")
+  end
+
+  it "makes distinction between the navigation text and label text of the links when a label is present" do
+    render_component(
+      previous_page: {
+        url: "previous-page",
+        title: "Previous page",
+        label: "1 of 3"
+      },
+      next_page: {
+        url: "next-page",
+        title: "Next page",
+        label: "2 of 3"
+      }
+    )
+
+    assert_select ".gem-c-pagination__item--previous .gem-c-pagination__link .visually-hidden", text: ":"
+    assert_select ".gem-c-pagination__item--next .gem-c-pagination__link .visually-hidden", text: ":"
+  end
+
+  it "does not make a distinction between the navigation text and label text of the links when labels are not present" do
+    render_component(
+      previous_page: {
+        url: "previous-page",
+        title: "Previous page"
+      },
+      next_page: {
+        url: "next-page",
+        title: "Next page"
+      }
+    )
+
+    assert_select ".gem-c-pagination__item--previous .gem-c-pagination__link .visually-hidden", false
+    assert_select ".gem-c-pagination__item--next .gem-c-pagination__link .visually-hidden", false
+  end
+
+  def assert_link(link)
+    assert_select "a[href=\"#{link}\"]"
+  end
+end
